# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple swagger for musiclib.",
        "title": "MusicLib",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
        "/v1/songs": {
            "get": {
                "description": "Gets song from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Get song",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id/group/name/date/text/link",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "int \u003e 0",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "int \u003e 0",
                        "name": "items",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/musiclib.Song"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "post": {
                "description": "Posts song to db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Post song",
                "parameters": [
                    {
                        "description": "Song JSON Object",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{\"group\":\"sample\", \"name\":\"mamedsd\", \"releaseDate\":\"2030-12-12\", \"text\":\"ss\", \"link\":\"sss\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        },
        "/v1/songs/{songId}": {
            "get": {
                "description": "Gets song from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Get song",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int \u003e= 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "int \u003e 0",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/musiclib.SongPaginated"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "delete": {
                "description": "deletes song from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Delete song",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "patch": {
                "description": "Patch song from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Patch song",
                "parameters": [
                    {
                        "description": "Song JSON Object",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{\"group\":\"Patched\", \"name\":\"PatchedName\", \"releaseDate\":\"2023-12-12\", \"text\":\"PatchedText\", \"link\":\"PatchedLink\"}"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        }
    },
    "definitions": {
        "musiclib.Song": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "musiclib.SongPaginated": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "text": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}